fix 自动备份增加完整半备份、自定义备份路径（待完善）

+++ web/action.py
@@ -5063,36 +5063,43 @@ class WebAction:
         return {"code": 0, "text": category_text}
 
     @staticmethod
-    def backup():
+    def backup(all_backup=False, bk_path=None):
+        """
+        @param all_backup  是否完整备份
+        @param bk_path     自定义备份路径
+        """
         try:
             # 创建备份文件夹
             config_path = Path(Config().get_config_path())
-            backup_file = f"bk_{time.strftime('%Y%m%d%H%M%S')}"
-            backup_path = config_path / "backup_file" / backup_file
+            backup_file = f"bk_{time.strftime('%Y%m%d%H%M%S')}" + "_all" if all_backup else ""
+            backup_path = bk_path or (config_path / "backup_file") / backup_file
             backup_path.mkdir(parents=True)
             # 把现有的相关文件进行copy备份
             shutil.copy(f'{config_path}/config.yaml', backup_path)
             shutil.copy(f'{config_path}/default-category.yaml', backup_path)
             shutil.copy(f'{config_path}/user.db', backup_path)
-            conn = sqlite3.connect(f'{backup_path}/user.db')
-            cursor = conn.cursor()
-            # 执行操作删除不需要备份的表
-            table_list = [
-                'SEARCH_RESULT_INFO',
-                'RSS_TORRENTS',
-                'DOUBAN_MEDIAS',
-                'TRANSFER_HISTORY',
-                'TRANSFER_UNKNOWN',
-                'TRANSFER_BLACKLIST',
-                'SYNC_HISTORY',
-                'DOWNLOAD_HISTORY',
-                'alembic_version'
-            ]
-            for table in table_list:
-                cursor.execute(f"""DROP TABLE IF EXISTS {table};""")
-            conn.commit()
-            cursor.close()
-            conn.close()
+
+            # 完整备份不删除表
+            if not all_backup:
+                conn = sqlite3.connect(f'{backup_path}/user.db')
+                cursor = conn.cursor()
+                # 执行操作删除不需要备份的表
+                table_list = [
+                    'SEARCH_RESULT_INFO',
+                    'RSS_TORRENTS',
+                    'DOUBAN_MEDIAS',
+                    'TRANSFER_HISTORY',
+                    'TRANSFER_UNKNOWN',
+                    'TRANSFER_BLACKLIST',
+                    'SYNC_HISTORY',
+                    'DOWNLOAD_HISTORY',
+                    'alembic_version'
+                ]
+                for table in table_list:
+                    cursor.execute(f"""DROP TABLE IF EXISTS {table};""")
+                conn.commit()
+                cursor.close()
+                conn.close()
             zip_file = str(backup_path) + '.zip'
             if os.path.exists(zip_file):
                 zip_file = str(backup_path) + '.zip'
